+++ types/assets/index.d.ts
@@ -12,8 +12,3 @@ declare module "*.py" {
 	const content: string;
 	export default content;
 }
-
-declare module "*.wasm?raw" {
-	const content: string;
-	export default content;
-}